Change API so that always calling release_string() when done with it is okay
For example, the file_get_name() function adds an extra reference to the
name before returning it, so the user can just call release_string() when
done with it.

This also allows to call file_get_name(), then close the file and reuse
the string afterwards. Because file_get_name() added an extra reference
the string is still vaild after the file was closed.

Also each function that takes a string object and stores it has to add
and extra reference to it, so the user can release the string once she's
done with it, without having to care if other objects are using it or not.