Merge lock and occupy handling
Rename the lock count to usage count and change it via the occupy/vacate
functions.

The lock/occupy handling is not recursive anymore. The list object occupies
its item when they are added to the list. Instead of locking them when the
list object was locked. This means that an object is locked while it is in
a list object.

Avoid the terms acquire and release.