The delivered of some callbacks should be delayed
Receiving the callback triggered by a getter before the getter response
is not critical in case of the file object API.

It is more critical in case of the process object API. Here the process
state changed callback could be received before the response of the spawn
function. This means that a client program receives a callback with a not
yet known.